kallsyms: Allow /proc/kallsyms to be disabled at compile time
If the kernel KASLR offset is leaked and this symbol data is available
to userspace, it can become a security issue as userspace possesses
information about the kernel's internals and knows which addresses to
target in memory in order to exploit the kernel.

Android does not use /proc/kallsyms at all, so this is safe and will not
break anything.

Test: disabled config option, /proc/kallsyms no longer exists
Signed-off-by: kdrag0n <dragon@khronodragon.com>

# Conflicts:
#	init/Kconfig
#	scripts/libufdt