cpu_input_boost: Clean up SchedTune boost handling
The previous method of tracking which SchedTune boosts were active using
extra bits in the global atomic state bitmask was inefficient as it
resulted in unnecessarily increased load on the atomic bitmask.
Additionally, atomic operations are more expensive than simple boolean
or integer operations.

Use separate booleans instead to alleviate the issue. This also comes with
the benefit of reducing the amount of arguments that must be passed to
the SchedTune boost helpers, as well as reducing calls to
get_boost_state.

Signed-off-by: kdrag0n <dragon@khronodragon.com>

# Conflicts:
#	drivers/cpufreq/cpu_input_boost.c
#	scripts/libufdt