[Squash]  simple_lmk: Introduce Simple LMK (Low Memory Killer) for Android
This is a replacement for the Android Low Memory Killer driver that simply
targets a predefined number of pages to be freed every time a memory
allocation attempt is about to utilize direct reclaim, or whenever it seems
like kswapd is having trouble freeing enough memory in a timely manner.
Processes are killed according to the priorities that Android gives them;
in other words, the least important processes are always killed first.

One way that memory reclaim is triggered is in the page-alloc slow path
after compaction has failed. When the Simple LMK memory reclaim process is
entered, memory is reclaimed in an infinite loop until the requested memory
allocation is fulfilled when pages are freed. During this process, Simple LMK
ensures that memory reclaim is invoked at least once every
CONFIG_ANDROID_SIMPLE_LMK_OOM_TIMEOUT milliseconds. OOM'd memory allocations
are placed on a queue and are prioritized over non-OOM'd memory allocations
when new pages are freed. This is to ensure that each OOM'd memory allocation
gets fulfilled as quickly as possible.

The other way that memory reclaim occurs is via kswapd. When kswapd runs
for longer than CONFIG_ANDROID_SIMPLE_LMK_KSWAPD_TIMEOUT milliseconds
without sleeping, then Simple LMK ensures that memory reclaim is invoked at
least once every CONFIG_ANDROID_SIMPLE_LMK_KSWAPD_TIMEOUT milliseconds.

*simple_lmk: Boost CPU and DDR bus when mem reclaim is invoked

Memory reclaim can be resource-intensive, which can lead to short periods
of lag right after a reclaim event takes place. To mitigate this, boost
the CPU and DDR bus to the max for a short period of time before doing
memory reclaim.

*simple_lmk: Only initialize driver after LMKD writes minfree settings

When booted in recovery, the OOM ADJ scores are not set according to what
they're expected to be in Android, which causes Simple LMK to misbehave.

In order to ensure that Simple LMK is only used with a compatible
userspace (i.e., Android), don't initialize the driver until LMKD in
Android attempts to write minfree settings. This is an indicator that
LMKD/Android is present and that Simple LMK can therefore be used.

*simple_lmk: Set WQ_MEM_RECLAIM bit

The WQ_MEM_RECLAIM bit should be set for workqueues which execute work
items that perform memory reclaim.

*simple_lmk: Clean up code style and update license header

Official Linux kernel style mandates function parameter alignment.

Also, update the license headers for the new style.

*simple_lmk: Fix potential race during init

If multiple processes attempt to write to the minfree parameter at the same
time, then bad things will happen. Use a tristate to avoid this.

*simple_lmk: Target older cached apps before killing newer ones

Enumerate the cached app ADJ rankings so that older apps with a higher ADJ are
killed first.

*simple_lmk: Use kswapd's actual execution time to determine when to kill

Using wall-time to determine how long kswapd has actually been running is
ambiguous since it doesn't account for how much of the time kswapd has been
preempted. As a result, when kswapd gets preempted a lot due to heavy CPU usage,
Simple LMK will take this as a hint to needlessly kill processes.

Make Simple LMK check kswapd's actual run-time to fix the spurious reclaims.

*simple_lmk: React faster to kswapd reclaim events

With the previous change, there would be up to a KSWAPD_LMK_EXPIRES delay before
Simple LMK would react to a kswapd reclaim being needed, and it would still make
it difficult for a kswapd reclaim to occur if an oom reclaim occurred less than
KSWAPD_LMK_EXPIRES ago.

To remedy Simple LMK's slow kswapd reaction time, poll every other jiffy once
kswapd has been running without sleeping for at least KSWAPD_LMK_EXPIRES' worth
of _wall time_. In addition, when a kswapd reclaim is needed, don't disregard it
if an oom reclaim has occurred recently; in other words, kswapd reclaims are
considered to be of high importance and should not be ignored.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>