simple_lmk: Improve victim selection and kill process
There are a few issues with Simple LMK's current victim selection
process, namely in that it just kills potential victims as it finds them
while holding an rcu read lock. It's more suitable to instead select the
largest task out of the potential victims for a specified adj range and
kill that instead, which is what this patch does.

Simple LMK now stores data on potential victims it finds in a large
array it has so that the array can be sorted later in descending order
of size. This makes Simple LMK kill as few processes as possible per
memory reclaim event. This additionally does away with killing the
victims while the rcu read lock is held for too long.

Furthermore, since Simple LMK uses a forced signal to kill victims,
there's no point in boosting the priority of a victim process after it's
been forced dead. The priority boost snakeoil is gone now as well.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: kdrag0n <dragon@khronodragon.com>

# Conflicts:
#	scripts/libufdt